- Patch hover button - add @shortcut in path - resize "nom" - split map and list - add date in table - add AUTO_INCREMENT in schema.sql - npm install before RunDevFrontend.run.xml - null support for sorting record - wrong date format in record - Add docs MySQL.md for Windows - Reorder styles.scss

+++ src/main/java/ummisco/gamaSenseIt/springServer/data/model/sensor/Parameter.java
@@ -70,7 +70,7 @@ public class Parameter {
     }
 
     public Parameter(long data, Date captureDate, ParameterMetadata parameterMetadata, Sensor sensor) {
-        this(ByteBuffer.allocate(Integer.BYTES).putLong(data).array(), captureDate, parameterMetadata, sensor);
+        this(ByteBuffer.allocate(Long.BYTES).putLong(data).array(), captureDate, parameterMetadata, sensor);
     }
 
     public Parameter(String data, Date captureDate, ParameterMetadata parameterMetadata, Sensor sensor) {
